requirements.txt
asgiref
attrs
coverage
dj-database-url
Django
django-cors-headers
django-environ
django-tenants
djangorestframework
iniconfig
kafka-python
packaging
Pillow
pluggy
psycopg2-binary
py
PyJWT
pyparsing
pytest
pytz
sqlparse
tomli
whitenoise
gunicorn